:card_file_box: Corrige las migraciones del nivel 6

+++ database/migrations/2022_08_01_041350_create_certificados_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
             $table->id();
             $table->date('fecha_emision');
             $table->string('num_registro', 25);
-            $table->mediumInteger('horas');
+            $table->smallInteger('horas'); // -32768 ~ 32767
             $table->foreignId('estudiante_id')->nullable()->constrained('estudiantes')
                 ->cascadeOnUpdate()
                 ->nullOnDelete();
